rename msg name
